+++ quadra/tasks/classification.py
@@ -36,7 +36,8 @@ from quadra.modules.classification import ClassificationModule
 from quadra.tasks.base import Evaluation, LightningTask, Task
 from quadra.trainers.classification import SklearnClassificationTrainer
 from quadra.utils import utils
-from quadra.utils.classification import get_results, save_classification_result
+from quadra.utils.classification import automatic_batch_size_computation, get_results, save_classification_result
+from quadra.utils.evaluation import automatic_datamodule_batch_size
 from quadra.utils.export import export_model, import_deployment_model
 from quadra.utils.models import get_feature, is_vision_transformer
 from quadra.utils.vit_explainability import VitAttentionGradRollout
@@ -97,6 +98,7 @@ class Classification(Generic[ClassificationDataModuleT], LightningTask[Classific
         self.export_folder: str = "deployment_model"
         self.deploy_info_file: str = "model.json"
         self.report_confmat: pd.DataFrame
+        self.best_model_path: Optional[str] = None
 
     @property
     def optimizer(self) -> torch.optim.Optimizer:
@@ -235,15 +237,23 @@ class Classification(Generic[ClassificationDataModuleT], LightningTask[Classific
         self.scheduler = self.config.scheduler
         self.module = self.config.model.module
 
+    def train(self):
+        """Train the model."""
+        super().train()
+        if (
+            self.trainer.checkpoint_callback is not None
+            and hasattr(self.trainer.checkpoint_callback, "best_model_path")
+            and self.trainer.checkpoint_callback.best_model_path is not None
+            and len(self.trainer.checkpoint_callback.best_model_path) > 0
+        ):
+            self.best_model_path = self.trainer.checkpoint_callback.best_model_path
+            log.info("Loading best epoch weights...")
+
     def test(self) -> None:
         """Test the model."""
         if not self.config.trainer.get("fast_dev_run"):
-            if self.trainer.checkpoint_callback is None:
-                raise ValueError("Checkpoint callback is not defined!")
             log.info("Starting testing!")
-            self.datamodule.setup(stage="test")
-            log.info("Using best epoch's weights for testing.")
-            self.trainer.test(datamodule=self.datamodule, model=self.module, ckpt_path="best")
+            self.trainer.test(datamodule=self.datamodule, model=self.module, ckpt_path=self.best_model_path)
 
     def export(self) -> None:
         """Generate deployment models for the task."""
@@ -255,20 +265,21 @@ class Classification(Generic[ClassificationDataModuleT], LightningTask[Classific
         else:
             idx_to_class = {v: k for k, v in self.datamodule.class_to_idx.items()}
 
-        if self.trainer.checkpoint_callback is None:
-            log.warning("No checkpoint callback found in the trainer, exporting the last model weights")
-        else:
-            best_model_path = self.trainer.checkpoint_callback.best_model_path  # type: ignore[attr-defined]
-            log.info("Saving deployment model for %s checkpoint", best_model_path)
+        # Get best model!
+        if self.best_model_path is not None:
+            log.info("Saving deployment model for %s checkpoint", self.best_model_path)
 
             module = self.module.load_from_checkpoint(
-                best_model_path,
+                self.best_model_path,
                 model=self.module.model,
                 optimizer=self.optimizer,
                 lr_scheduler=self.scheduler,
                 criterion=self.module.criterion,
                 gradcam=False,
             )
+        else:
+            log.warning("No checkpoint callback found in the trainer, exporting the last model weights")
+            module = self.module
 
         input_shapes = self.config.export.input_shapes
 
@@ -303,9 +314,9 @@ class Classification(Generic[ClassificationDataModuleT], LightningTask[Classific
         log.info("Generating report!")
         if not self.run_test or self.config.trainer.get("fast_dev_run"):
             self.datamodule.setup(stage="test")
-        best_model_path = self.trainer.checkpoint_callback.best_model_path  # type: ignore[union-attr]
+
         predictions_outputs = self.trainer.predict(
-            model=self.module, datamodule=self.datamodule, ckpt_path=best_model_path
+            model=self.module, datamodule=self.datamodule, ckpt_path=self.best_model_path
         )
         if not predictions_outputs:
             log.warning("There is no prediction to generate the report. Skipping report generation.")
@@ -420,6 +431,7 @@ class SklearnClassification(Generic[SklearnClassificationDataModuleT], Task[Skle
         config: The experiment configuration
         device: The device to use. Defaults to None.
         output: Dictionary defining which kind of outputs to generate. Defaults to None.
+        automatic_batch_size: Whether to automatically find the largest batch size that fits in memory.
     """
 
     def __init__(
@@ -427,6 +439,7 @@ class SklearnClassification(Generic[SklearnClassificationDataModuleT], Task[Skle
         config: DictConfig,
         output: DictConfig,
         device: str,
+        automatic_batch_size: DictConfig,
     ):
         super().__init__(config=config)
 
@@ -445,6 +458,7 @@ class SklearnClassification(Generic[SklearnClassificationDataModuleT], Task[Skle
         self.deploy_info_file = "model.json"
         self.train_dataloader_list: List[torch.utils.data.DataLoader] = []
         self.test_dataloader_list: List[torch.utils.data.DataLoader] = []
+        self.automatic_batch_size = automatic_batch_size
 
     @property
     def device(self) -> str:
@@ -460,6 +474,13 @@ class SklearnClassification(Generic[SklearnClassificationDataModuleT], Task[Skle
         # prepare_data() must be explicitly called if the task does not include a lightining training
         self.datamodule.prepare_data()
         self.datamodule.setup(stage="fit")
+
+        if not self.automatic_batch_size.disable and self.device != "cpu":
+            self.datamodule.batch_size = automatic_batch_size_computation(
+                datamodule=self.datamodule,
+                backbone=self.backbone,
+                starting_batch_size=self.automatic_batch_size.starting_batch_size,
+            )
 
         self.train_dataloader_list = list(self.datamodule.train_dataloader())
         self.test_dataloader_list = list(self.datamodule.val_dataloader())
@@ -759,8 +780,6 @@ class SklearnTestClassification(Evaluation[SklearnClassificationDataModuleT]):
         self.datamodule.prepare_data()
         self.datamodule.setup(stage="test")
 
-        self.test_dataloader = self.datamodule.test_dataloader()
-
         # Configure trainer
         self.trainer = self.config.trainer
 
@@ -836,8 +855,11 @@ class SklearnTestClassification(Evaluation[SklearnClassificationDataModuleT]):
         trainer = hydra.utils.instantiate(trainer_config, backbone=self.backbone, classifier=self.classifier)
         self._trainer = trainer
 
+    @automatic_datamodule_batch_size(batch_size_attribute_name="batch_size")
     def test(self) -> None:
         """Run the test."""
+        self.test_dataloader = self.datamodule.test_dataloader()
+
         _, pd_cm, accuracy, res, cams = self.trainer.test(
             test_dataloader=self.test_dataloader,
             idx_to_class=self.idx_to_class,
@@ -933,7 +955,7 @@ class ClassificationEvaluation(Evaluation[ClassificationDataModuleT]):
         if "classifier" in model_config:
             log.info("Instantiating classifier <%s>", model_config.classifier["_target_"])
             return hydra.utils.instantiate(
-                model_config.classifier, out_features=self.datamodule.num_classes, _convert_="partial"
+                model_config.classifier, out_features=len(self.model_data["classes"]), _convert_="partial"
             )
 
         raise ValueError("A `classifier` definition must be specified in the config")
@@ -947,13 +969,6 @@ class ClassificationEvaluation(Evaluation[ClassificationDataModuleT]):
     def deployment_model(self, model_path: str):
         """Set the deployment model."""
         file_extension = os.path.splitext(model_path)[1]
-
-        if "classes" in self.model_data:
-            self.datamodule.class_to_idx = {v: int(k) for k, v in self.model_data["classes"].items()}
-            self.datamodule.num_classes = len(self.datamodule.class_to_idx)
-        else:
-            raise ValueError("Field 'classes' is missing from json's model_data")
-
         model_architecture = None
         if file_extension == ".pth":
             model_config = OmegaConf.load(os.path.join(Path(model_path).parent, "model_config.yaml"))
@@ -976,8 +991,14 @@ class ClassificationEvaluation(Evaluation[ClassificationDataModuleT]):
 
     def prepare(self) -> None:
         """Prepare the evaluation."""
-        self.datamodule = self.config.datamodule
         super().prepare()
+        self.datamodule = self.config.datamodule
+        self.datamodule.class_to_idx = {v: int(k) for k, v in self.model_data["classes"].items()}
+        self.datamodule.num_classes = len(self.datamodule.class_to_idx)
+
+        # prepare_data() must be explicitly called because there is no training
+        self.datamodule.prepare_data()
+        self.datamodule.setup(stage="test")
 
     def prepare_gradcam(self) -> None:
         """Initializing gradcam for the predictions."""
@@ -1005,12 +1026,10 @@ class ClassificationEvaluation(Evaluation[ClassificationDataModuleT]):
             log.warning("Gradcam not implemented for this backbone, it will not be computed")
             self.gradcam = False
 
+    @automatic_datamodule_batch_size(batch_size_attribute_name="batch_size")
     def test(self) -> None:
         """Perform test."""
         log.info("Running test")
-        # prepare_data() must be explicitly called because there is no training
-        self.datamodule.prepare_data()
-        self.datamodule.setup(stage="test")
         test_dataloader = self.datamodule.test_dataloader()
 
         image_labels = []